What a mess.
IE 9 doesn't do gradients without SVG, and that's been a hassle to figure out
how to get working. So I've opted to hand-code in the Microsoft filter for a
gradient. It's not radial, just darkens the top of the image.

Of course, to get this to work, you have to also redeclare the backgrond image
by itself. THEN, for all the other browsers, we'll do like we did before.